+++ CHANGELOG.md
@@ -5,6 +5,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added access token authentication
+- Added User Managed Identity auth - Needs Testing
+- Added System Manged Identity auth - Needs Testing
+- Added Environment variable auth - Needs Testing
+
+### Changed
+
+- Updated authentication logic to support new authentication methods and parameters in `Get-EmMdmGraphAuth`.
+
 ## [0.1.3] - 2024-07-28
 
 ### Fix